manager streaming

+++ services/repository-managers/src/main/java/org/sagebionetworks/repo/manager/migration/BackupFileStream.java
@@ -42,5 +42,5 @@ public interface BackupFileStream {
 	 * @throws IOException
 	 */
 	public void writeBackupFile(OutputStream out, Iterable<MigratableDatabaseObject<?,?>> stream, BackupAliasType backupAliasType,
-			int maximumRowsPerFile) throws IOException;
+			long maximumRowsPerFile) throws IOException;
 }
